refactor: 型定義ファイル・ディレクトリの命名を変更
