+++ components/CheckinDetail.component.tsx
@@ -7,7 +7,7 @@ import { useDate } from '../hooks/useDate'
 import { useUtils } from '../hooks/useUtils'
 import colors from '../constants/Colors'
 import { useFoursquare } from '../hooks/useFoursquare'
-import { Checkin, User } from '../interface/Foursquare.type'
+import { Checkin, User } from '../types/Foursquare'
 import { commonStyles } from '../styles/styles'
 import { ImageCarousel } from '../components/carousel/ImageCarousel.component'
 import CoinIcon from './CoinIcon.component'
