+++ components/card/checkin.component.tsx
@@ -4,7 +4,7 @@ import { Avatar, Image, Icon } from 'react-native-elements'
 import { ScrollView } from 'react-native-gesture-handler'
 import ImageViewer from 'react-native-image-zoom-viewer'
 import { useNavigation } from '@react-navigation/core'
-import type { Checkin } from '../../interface/Foursquare.type'
+import type { Checkin } from '../../types/Foursquare'
 import window from '../../constants/Layout'
 import { useDate } from '../../hooks/useDate'
 import { useUtils } from '../../hooks/useUtils'
