+++ screens/CheckinCalendar.tsx
@@ -4,7 +4,7 @@ import { Agenda, DateObject } from 'react-native-calendars'
 import { useDate } from '../hooks/useDate'
 import { useFoursquare } from '../hooks/useFoursquare'
 import { useUtils } from '../hooks/useUtils'
-import type { Checkin } from '../interface/Foursquare.type'
+import type { Checkin } from '../types/Foursquare'
 import { Timeline } from '../components/Timeline.component'
 import { useRecoil } from '../hooks/useRecoil'
 import Colors from '../constants/Colors'
