+++ types/Foursquare.ts
